removed nav button as the already exist , added user interface, set up user data service, set up user service set up api call

+++ ParkMEHandler/src/main/java/com/example/ParkMEHandler/User.java
@@ -32,8 +32,8 @@ public class User {
     }
 
     // Getters
-    public Long getUserId() {
-        return userId;
+    public Integer getUserId() {
+        return Math.toIntExact(userId);
     }
 
     public String getEmail() {
